+++ pyrad/__init__.py
@@ -40,7 +40,7 @@ __docformat__ = 'epytext en'
 
 __author__ = 'Christian Giese <developer@gicnet.de>'
 __url__ = 'http://pyrad.readthedocs.io/en/latest/?badge=latest'
-__copyright__ = 'Copyright 2002-2019 Wichert Akkerman and Christian Giese. All rights reserved.'
-__version__ = '2.2'
+__copyright__ = 'Copyright 2002-2020 Wichert Akkerman and Christian Giese. All rights reserved.'
+__version__ = '2.3'
 
 __all__ = ['client', 'dictionary', 'packet', 'server', 'tools', 'dictfile']